refactor(schema): split core into modules
