+++ src/schema/core.rs
@@ -64,15 +64,15 @@ pub enum SchemaState {
 /// previously split across SchemaManager and SchemaInterpreter.
 pub struct SchemaCore {
     /// Thread-safe storage for loaded schemas
-    schemas: Mutex<HashMap<String, Schema>>,
+    pub(crate) schemas: Mutex<HashMap<String, Schema>>,
     /// All schemas known to the system and their load state
-    available: Mutex<HashMap<String, (Schema, SchemaState)>>,
+    pub(crate) available: Mutex<HashMap<String, (Schema, SchemaState)>>,
     /// Unified database operations (required)
-    db_ops: std::sync::Arc<crate::db_operations::DbOperations>,
+    pub(crate) db_ops: std::sync::Arc<crate::db_operations::DbOperations>,
     /// Schema directory path
-    schemas_dir: PathBuf,
+    pub(crate) schemas_dir: PathBuf,
     /// Message bus for event-driven communication
-    message_bus: Arc<MessageBus>,
+    pub(crate) message_bus: Arc<MessageBus>,
 }
 
 impl SchemaCore {
@@ -121,125 +121,6 @@ impl SchemaCore {
         Self::new_for_testing("data")
     }
 
-    /// Persist all schema load states using DbOperations
-    fn persist_states(&self) -> Result<(), SchemaError> {
-        let available = self
-            .available
-            .lock()
-            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
-
-        for (name, (_, state)) in available.iter() {
-            self.db_ops.store_schema_state(name, *state)?;
-        }
-
-        Ok(())
-    }
-
-    /// Load schema states using DbOperations
-    pub fn load_states(&self) -> HashMap<String, SchemaState> {
-        self.db_ops.get_all_schema_states().unwrap_or_default()
-    }
-
-    /// Persists a schema using DbOperations
-    fn persist_schema(&self, schema: &Schema) -> Result<(), SchemaError> {
-        self.db_ops.store_schema(&schema.name, schema)
-    }
-
-    fn fix_transform_outputs(&self, schema: &mut Schema) {
-        for (field_name, field) in schema.fields.iter_mut() {
-            if let Some(transform) = field.transform() {
-                let out_schema = transform.get_output();
-                if out_schema.starts_with("test.") {
-                    let mut new_transform = (*transform).clone();
-                    new_transform.set_output(format!("{}.{}", schema.name, field_name));
-                    field.set_transform(new_transform);
-                }
-            }
-        }
-    }
-
-    /// Auto-register field transforms with TransformManager during schema loading
-    fn register_schema_transforms(&self, schema: &Schema) -> Result<(), SchemaError> {
-        info!("🔧 DEBUG: Auto-registering transforms for schema: {}", schema.name);
-        info!("🔍 DEBUG: Schema has {} fields to check for transforms", schema.fields.len());
-        
-        for (field_name, field) in &schema.fields {
-            info!("🔍 DEBUG: Checking field '{}.{}' for transforms", schema.name, field_name);
-            if let Some(transform) = field.transform() {
-                info!(
-                    "📋 Found transform on field {}.{}: inputs={:?}, logic={}, output={}",
-                    schema.name, field_name, transform.get_inputs(), transform.logic, transform.get_output()
-                );
-                
-                let transform_id = format!("{}.{}", schema.name, field_name);
-                
-                // Store the transform in the database so it can be loaded by TransformManager
-                if let Err(e) = self.db_ops.store_transform(&transform_id, transform) {
-                    log::error!(
-                        "Failed to store transform {}: {}",
-                        transform_id, e
-                    );
-                    continue;
-                }
-                
-                info!("✅ Stored transform {} for auto-registration", transform_id);
-                
-                // 🛠️ FIX: Create field-to-transform mappings for TransformOrchestrator
-                // This is the missing piece - we need to map each input field to this transform
-                for input_field in transform.get_inputs() {
-                    info!(
-                        "🔗 Creating field mapping: '{}' → '{}' transform",
-                        input_field, transform_id
-                    );
-                    
-                    // Store field mapping in database for TransformManager to load
-                    if let Err(e) = self.store_field_to_transform_mapping(input_field, &transform_id) {
-                        log::error!(
-                            "Failed to store field mapping '{}' → '{}': {}",
-                            input_field, transform_id, e
-                        );
-                    } else {
-                        info!(
-                            "✅ Stored field mapping: '{}' → '{}' transform",
-                            input_field, transform_id
-                        );
-                    }
-                }
-            }
-        }
-        
-        Ok(())
-    }
-
-    /// Store field-to-transform mapping in database for TransformManager to load
-    fn store_field_to_transform_mapping(&self, field_key: &str, transform_id: &str) -> Result<(), SchemaError> {
-        // Use the same key format as TransformManager
-        const FIELD_TO_TRANSFORMS_KEY: &str = "map_field_to_transforms";
-        
-        // Load existing mappings using the correct method
-        let mut field_mappings: std::collections::HashMap<String, std::collections::HashSet<String>> =
-            if let Some(data) = self.db_ops.get_transform_mapping(FIELD_TO_TRANSFORMS_KEY)? {
-                serde_json::from_slice(&data).unwrap_or_default()
-            } else {
-                std::collections::HashMap::new()
-            };
-        
-        // Add this mapping
-        field_mappings
-            .entry(field_key.to_string())
-            .or_default()
-            .insert(transform_id.to_string());
-        
-        // Store updated mappings using the correct method
-        let json = serde_json::to_vec(&field_mappings).map_err(|e| {
-            SchemaError::InvalidData(format!("Failed to serialize field mappings: {}", e))
-        })?;
-        self.db_ops.store_transform_mapping(FIELD_TO_TRANSFORMS_KEY, &json)?;
-        
-        info!("💾 Updated field mappings in database: {} fields mapped", field_mappings.len());
-        
-        Ok(())
-    }
 
     /// Load a schema into memory and persist it to disk.
     /// This preserves existing schema state if it exists, otherwise defaults to Available.
@@ -538,287 +419,6 @@ impl SchemaCore {
         Ok(schemas)
     }
 
-    /// Discover schemas from the schemas directory
-    pub fn discover_schemas_from_files(&self) -> Result<Vec<Schema>, SchemaError> {
-        let mut discovered_schemas = Vec::new();
-
-        info!("Discovering schemas from {}", self.schemas_dir.display());
-        if let Ok(entries) = std::fs::read_dir(&self.schemas_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.extension().map(|e| e == "json").unwrap_or(false) {
-                    if let Ok(contents) = std::fs::read_to_string(&path) {
-                        let mut schema_opt = serde_json::from_str::<Schema>(&contents).ok();
-                        if schema_opt.is_none() {
-                            if let Ok(json_schema) =
-                                serde_json::from_str::<JsonSchemaDefinition>(&contents)
-                            {
-                                if let Ok(schema) = self.interpret_schema(json_schema) {
-                                    schema_opt = Some(schema);
-                                }
-                            }
-                        }
-                        if let Some(mut schema) = schema_opt {
-                            self.fix_transform_outputs(&mut schema);
-                            let schema_name = schema.name.clone();
-                            discovered_schemas.push(schema);
-                            info!("Discovered schema '{}' from file", schema_name);
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(discovered_schemas)
-    }
-
-    /// Discover schemas from the available_schemas directory
-    pub fn discover_available_schemas(&self) -> Result<Vec<Schema>, SchemaError> {
-        info!("🔍 DEBUG: Starting discovery from available_schemas directory");
-        let mut discovered_schemas = Vec::new();
-        let available_schemas_dir = PathBuf::from("available_schemas");
-
-        info!(
-            "Discovering available schemas from {}",
-            available_schemas_dir.display()
-        );
-        if let Ok(entries) = std::fs::read_dir(&available_schemas_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.extension().map(|e| e == "json").unwrap_or(false) {
-                    if let Ok(contents) = std::fs::read_to_string(&path) {
-                        let mut schema_opt = serde_json::from_str::<Schema>(&contents).ok();
-                        if schema_opt.is_none() {
-                            if let Ok(json_schema) =
-                                serde_json::from_str::<JsonSchemaDefinition>(&contents)
-                            {
-                                if let Ok(schema) = self.interpret_schema(json_schema) {
-                                    schema_opt = Some(schema);
-                                }
-                            }
-                        }
-                        if let Some(mut schema) = schema_opt {
-                            self.fix_transform_outputs(&mut schema);
-                            let schema_name = schema.name.clone();
-                            discovered_schemas.push(schema);
-                            info!("Discovered available schema '{}' from file", schema_name);
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(discovered_schemas)
-    }
-
-    /// Load all schemas from the available_schemas directory into SchemaCore
-    pub fn load_available_schemas_from_directory(&self) -> Result<(), SchemaError> {
-        let discovered_schemas = self.discover_available_schemas()?;
-
-        for schema in discovered_schemas {
-            let schema_name = schema.name.clone();
-            info!("Loading available schema '{}' into SchemaCore", schema_name);
-            self.load_schema_internal(schema)?;
-        }
-
-        info!(
-            "Loaded {} schemas from available_schemas directory",
-            self.list_available_schemas()?.len()
-        );
-        Ok(())
-    }
-
-    // ========== UNIFIED SCHEMA DISCOVERY API ==========
-
-    /// Single entry point for all schema discovery and loading
-    /// Consolidates all existing discovery methods (no sample manager)
-    pub fn discover_and_load_all_schemas(&self) -> Result<SchemaLoadingReport, SchemaError> {
-        info!("🔍 Starting unified schema discovery and loading");
-
-        let mut discovered_schemas = Vec::new();
-        let mut failed_schemas = Vec::new();
-        let mut loading_sources = HashMap::new();
-
-        // Get current schemas in memory to avoid unnecessary reloading
-        let current_schemas = {
-            let available = self.available.lock().map_err(|_| {
-                SchemaError::InvalidData("Failed to acquire schema lock".to_string())
-            })?;
-            available
-                .keys()
-                .cloned()
-                .collect::<std::collections::HashSet<String>>()
-        };
-
-        // 1. FIRST: Load existing schema states from sled persistence
-        info!("📋 Loading existing schema states from persistence first");
-        let schema_states = self.load_states();
-        for schema_name in schema_states.keys() {
-            loading_sources.insert(schema_name.clone(), SchemaSource::Persistence);
-            info!("Loaded persisted schema state for '{}'", schema_name);
-        }
-
-        // 2. SECOND: Discover from available_schemas/ directory (without overwriting states)
-        match self.discover_available_schemas() {
-            Ok(schemas) => {
-                for schema in schemas {
-                    let schema_name = schema.name.clone();
-                    discovered_schemas.push(schema_name.clone());
-                    
-                    // Only update loading source if not already loaded from persistence
-                    if !loading_sources.contains_key(&schema_name) {
-                        loading_sources.insert(schema_name.clone(), SchemaSource::AvailableDirectory);
-                    }
-
-                    // Only load if not already in memory
-                    if !current_schemas.contains(&schema_name) {
-                        info!(
-                            "Loading new schema '{}' from available_schemas/ (preserving persisted state)",
-                            schema_name
-                        );
-                        if let Err(e) = self.load_schema_internal(schema) {
-                            failed_schemas.push((schema_name, e.to_string()));
-                        }
-                    } else {
-                        info!(
-                            "Schema '{}' already in memory, skipping reload",
-                            schema_name
-                        );
-                    }
-                }
-            }
-            Err(e) => {
-                info!("Failed to discover schemas from available_schemas/: {}", e);
-            }
-        }
-
-        // 3. THIRD: Discover from data/schemas/ directory (without overwriting states)
-        match self.discover_schemas_from_files() {
-            Ok(schemas) => {
-                for schema in schemas {
-                    let schema_name = schema.name.clone();
-                    if !discovered_schemas.contains(&schema_name) {
-                        discovered_schemas.push(schema_name.clone());
-                        
-                        // Only update loading source if not already loaded from persistence
-                        if !loading_sources.contains_key(&schema_name) {
-                            loading_sources.insert(schema_name.clone(), SchemaSource::DataDirectory);
-                        }
-
-                        // Only load if not already in memory
-                        if !current_schemas.contains(&schema_name) {
-                            info!("Loading new schema '{}' from data/schemas/ (preserving persisted state)", schema_name);
-                            if let Err(e) = self.load_schema_internal(schema) {
-                                failed_schemas.push((schema_name, e.to_string()));
-                            }
-                        } else {
-                            info!(
-                                "Schema '{}' already in memory, skipping reload",
-                                schema_name
-                            );
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                info!("Failed to discover schemas from data/schemas/: {}", e);
-            }
-        }
-
-        // 4. Get loaded schemas (approved state)
-        let loaded_schemas = self
-            .list_schemas_by_state(SchemaState::Approved)
-            .unwrap_or_else(|_| Vec::new());
-
-        info!(
-            "✅ Schema discovery complete: {} discovered, {} loaded, {} failed",
-            discovered_schemas.len(),
-            loaded_schemas.len(),
-            failed_schemas.len()
-        );
-
-        Ok(SchemaLoadingReport {
-            discovered_schemas,
-            loaded_schemas,
-            failed_schemas,
-            schema_states,
-            loading_sources,
-            last_updated: chrono::Utc::now(),
-        })
-    }
-
-    /// Initialize schema system - called during node startup
-    pub fn initialize_schema_system(&self) -> Result<(), SchemaError> {
-        info!("🚀 Initializing schema system");
-        self.discover_and_load_all_schemas()?;
-        info!("✅ Schema system initialized successfully");
-        Ok(())
-    }
-
-    /// Get comprehensive schema status for UI
-    pub fn get_schema_status(&self) -> Result<SchemaLoadingReport, SchemaError> {
-        info!("📊 Getting schema status");
-
-        let schema_states = self.load_states();
-        let loaded_schemas = self
-            .list_schemas_by_state(SchemaState::Approved)
-            .unwrap_or_else(|_| Vec::new());
-
-        // Get all known schemas from states
-        let discovered_schemas: Vec<String> = schema_states.keys().cloned().collect();
-
-        // Create loading sources map (we don't track this in current implementation)
-        let loading_sources: HashMap<String, SchemaSource> = discovered_schemas
-            .iter()
-            .map(|name| (name.clone(), SchemaSource::Persistence))
-            .collect();
-
-        Ok(SchemaLoadingReport {
-            discovered_schemas,
-            loaded_schemas,
-            failed_schemas: Vec::new(), // No failures in status check
-            schema_states,
-            loading_sources,
-            last_updated: chrono::Utc::now(),
-        })
-    }
-
-    // ========== LEGACY CONSOLIDATED SCHEMA API ==========
-
-    /// Fetch available schemas from files (both data/schemas and available_schemas directories)
-    /// DEPRECATED: Use discover_and_load_all_schemas() instead
-    pub fn fetch_available_schemas(&self) -> Result<Vec<String>, SchemaError> {
-        let mut all_schemas = Vec::new();
-
-        // Get schemas from the default data/schemas directory
-        let discovered_default = self.discover_schemas_from_files()?;
-        all_schemas.extend(discovered_default.into_iter().map(|s| s.name));
-
-        // Get schemas from the available_schemas directory
-        let discovered_available = self.discover_available_schemas()?;
-        all_schemas.extend(discovered_available.into_iter().map(|s| s.name));
-
-        // Remove duplicates while preserving order
-        let mut unique_schemas = Vec::new();
-        for schema_name in all_schemas {
-            if !unique_schemas.contains(&schema_name) {
-                unique_schemas.push(schema_name);
-            }
-        }
-
-        Ok(unique_schemas)
-    }
-
-    /// Load schema state from sled
-    pub fn load_schema_state(&self) -> Result<HashMap<String, SchemaState>, SchemaError> {
-        let states = self.load_states();
-        Ok(states)
-    }
-
-    /// Load available schemas from sled and files
-    pub fn load_available_schemas(&self) -> Result<(), SchemaError> {
-        self.load_schemas_from_disk()
-    }
 
     /// Check if a schema can be queried (must be Approved)
     pub fn can_query_schema(&self, schema_name: &str) -> bool {
@@ -1193,215 +793,6 @@ impl SchemaCore {
         drop(schemas);
         self.set_schema_state(schema_name, SchemaState::Available)?;
         info!("Schema '{}' unloaded and marked as Available", schema_name);
-        Ok(())
-    }
-
-    /// Loads all schema files from both the schemas directory and available_schemas directory and restores their states.
-    /// Schemas marked as Approved will be loaded into active memory.
-    pub fn load_schemas_from_disk(&self) -> Result<(), SchemaError> {
-        let states = self.load_states();
-
-        // Load from default schemas directory
-        info!("Loading schemas from {}", self.schemas_dir.display());
-        self.load_schemas_from_directory(&self.schemas_dir, &states)?;
-
-        // Load from available_schemas directory
-        let available_schemas_dir = PathBuf::from("available_schemas");
-        info!("Loading schemas from {}", available_schemas_dir.display());
-        self.load_schemas_from_directory(&available_schemas_dir, &states)?;
-
-        // Persist any changes to schema states from newly discovered schemas
-        self.persist_states()?;
-
-        Ok(())
-    }
-
-    /// Helper method to load schemas from a specific directory
-    fn load_schemas_from_directory(
-        &self,
-        dir: &PathBuf,
-        states: &HashMap<String, SchemaState>,
-    ) -> Result<(), SchemaError> {
-        if let Ok(entries) = std::fs::read_dir(dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.extension().map(|e| e == "json").unwrap_or(false) {
-                    if let Ok(contents) = std::fs::read_to_string(&path) {
-                        let mut schema_opt = serde_json::from_str::<Schema>(&contents).ok();
-                        if schema_opt.is_none() {
-                            if let Ok(json_schema) =
-                                serde_json::from_str::<JsonSchemaDefinition>(&contents)
-                            {
-                                if let Ok(schema) = self.interpret_schema(json_schema) {
-                                    schema_opt = Some(schema);
-                                }
-                            }
-                        }
-                        if let Some(mut schema) = schema_opt {
-                            self.fix_transform_outputs(&mut schema);
-                            let name = schema.name.clone();
-                            let state =
-                                states.get(&name).copied().unwrap_or(SchemaState::Available);
-                            {
-                                let mut available = self.available.lock().map_err(|_| {
-                                    SchemaError::InvalidData(
-                                        "Failed to acquire schema lock".to_string(),
-                                    )
-                                })?;
-                                available.insert(name.clone(), (schema.clone(), state));
-                            }
-                            if state == SchemaState::Approved {
-                                let mut loaded = self.schemas.lock().map_err(|_| {
-                                    SchemaError::InvalidData(
-                                        "Failed to acquire schema lock".to_string(),
-                                    )
-                                })?;
-                                loaded.insert(name.clone(), schema);
-                                drop(loaded); // Release the lock before calling map_fields
-
-                                // Ensure fields have proper ARefs assigned
-                                let _ = self.map_fields(&name);
-                            }
-                            info!(
-                                "Loaded schema '{}' from {} with state: {:?}",
-                                name,
-                                dir.display(),
-                                state
-                            );
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Loads schema states from sled and loads schemas that are marked as Approved.
-    pub fn load_schema_states_from_disk(&self) -> Result<(), SchemaError> {
-        let states = self.load_states();
-        info!("Loading schema states from sled: {:?}", states);
-        info!(
-            "DEBUG: load_schema_states_from_disk called with {} states",
-            states.len()
-        );
-        let mut available = self
-            .available
-            .lock()
-            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
-        let mut schemas = self
-            .schemas
-            .lock()
-            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
-
-        for (name, state) in states {
-            info!("DEBUG: Processing schema '{}' with state {:?}", name, state);
-            if state == SchemaState::Approved {
-                // Load the actual schema from sled database into active memory
-                match self.db_ops.get_schema(&name) {
-                    Ok(Some(mut schema)) => {
-                        info!(
-                            "Auto-loading approved schema '{}' from sled with {} fields: {:?}",
-                            name,
-                            schema.fields.len(),
-                            schema.fields.keys().collect::<Vec<_>>()
-                        );
-
-                        // 🔄 Log ref_atom_uuid values during schema loading
-                        info!(
-                            "🔄 SCHEMA_LOAD - Loading schema '{}' with {} fields",
-                            name,
-                            schema.fields.len()
-                        );
-                        for (field_name, field_def) in &schema.fields {
-                            use crate::schema::types::Field;
-                            match field_def.ref_atom_uuid() {
-                                Some(uuid) => info!(
-                                    "📋 Field {}.{} has ref_atom_uuid: {}",
-                                    name, field_name, uuid
-                                ),
-                                None => info!(
-                                    "📋 Field {}.{} has ref_atom_uuid: None",
-                                    name, field_name
-                                ),
-                            }
-                        }
-
-                        self.fix_transform_outputs(&mut schema);
-                        info!("After fix_transform_outputs, auto-loaded schema '{}' has {} fields: {:?}", name, schema.fields.len(), schema.fields.keys().collect::<Vec<_>>());
-                        schemas.insert(name.clone(), schema.clone());
-                        available.insert(name.clone(), (schema, state));
-                        drop(schemas); // Release the lock before calling map_fields
-                        drop(available); // Release the lock before calling map_fields
-
-                        // Ensure fields have proper ARefs assigned
-                        let _ = self.map_fields(&name);
-
-                        // Re-acquire locks for the next iteration
-                        available = self.available.lock().map_err(|_| {
-                            SchemaError::InvalidData("Failed to acquire schema lock".to_string())
-                        })?;
-                        schemas = self.schemas.lock().map_err(|_| {
-                            SchemaError::InvalidData("Failed to acquire schema lock".to_string())
-                        })?;
-                    }
-                    Ok(None) => {
-                        info!("Schema '{}' not found in sled, creating empty schema", name);
-                        available.insert(name.clone(), (Schema::new(name), SchemaState::Available));
-                    }
-                    Err(e) => {
-                        info!("Failed to load schema '{}' from sled: {}", name, e);
-                        available.insert(name.clone(), (Schema::new(name), SchemaState::Available));
-                    }
-                }
-            } else {
-                // Load the actual schema from sled for non-Approved states too
-                match self.db_ops.get_schema(&name) {
-                    Ok(Some(mut schema)) => {
-                        // 🔄 Log ref_atom_uuid values during schema loading (non-Approved)
-                        info!(
-                            "🔄 SCHEMA_LOAD - Loading schema '{}' (state: {:?}) with {} fields",
-                            name,
-                            state,
-                            schema.fields.len()
-                        );
-                        for (field_name, field_def) in &schema.fields {
-                            use crate::schema::types::Field;
-                            match field_def.ref_atom_uuid() {
-                                Some(uuid) => info!(
-                                    "📋 Field {}.{} has ref_atom_uuid: {}",
-                                    name, field_name, uuid
-                                ),
-                                None => info!(
-                                    "📋 Field {}.{} has ref_atom_uuid: None",
-                                    name, field_name
-                                ),
-                            }
-                        }
-
-                        self.fix_transform_outputs(&mut schema);
-                        info!(
-                            "Loading schema '{}' from sled with state {:?} and {} fields: {:?}",
-                            name,
-                            state,
-                            schema.fields.len(),
-                            schema.fields.keys().collect::<Vec<_>>()
-                        );
-                        available.insert(name.clone(), (schema, state));
-                    }
-                    Ok(None) => {
-                        info!("Schema '{}' not found in sled, creating empty schema", name);
-                        available.insert(name.clone(), (Schema::new(name), state));
-                    }
-                    Err(e) => {
-                        info!(
-                            "Failed to load schema '{}' from sled: {}, creating empty schema",
-                            name, e
-                        );
-                        available.insert(name.clone(), (Schema::new(name), state));
-                    }
-                }
-            }
-        }
         Ok(())
     }
