+++ src/schema/persistence.rs
@@ -1,0 +1,245 @@
+use super::{SchemaCore, SchemaState};
+use crate::schema::types::{JsonSchemaDefinition, Schema, SchemaError};
+use log::info;
+use std::collections::HashMap;
+use std::path::PathBuf;
+
+impl SchemaCore {
+    /// Persist all schema load states using DbOperations
+    pub(crate) fn persist_states(&self) -> Result<(), SchemaError> {
+        let available = self
+            .available
+            .lock()
+            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
+
+        for (name, (_, state)) in available.iter() {
+            self.db_ops.store_schema_state(name, *state)?;
+        }
+
+        Ok(())
+    }
+
+    /// Load schema states using DbOperations
+    pub fn load_states(&self) -> HashMap<String, SchemaState> {
+        self.db_ops.get_all_schema_states().unwrap_or_default()
+    }
+
+    /// Persists a schema using DbOperations
+    pub(crate) fn persist_schema(&self, schema: &Schema) -> Result<(), SchemaError> {
+        self.db_ops.store_schema(&schema.name, schema)
+    }
+
+    /// Loads all schema files from both the schemas directory and available_schemas directory and restores their states.
+    /// Schemas marked as Approved will be loaded into active memory.
+    pub fn load_schemas_from_disk(&self) -> Result<(), SchemaError> {
+        let states = self.load_states();
+
+        // Load from default schemas directory
+        info!("Loading schemas from {}", self.schemas_dir.display());
+        self.load_schemas_from_directory(&self.schemas_dir, &states)?;
+
+        // Load from available_schemas directory
+        let available_schemas_dir = PathBuf::from("available_schemas");
+        info!("Loading schemas from {}", available_schemas_dir.display());
+        self.load_schemas_from_directory(&available_schemas_dir, &states)?;
+
+        // Persist any changes to schema states from newly discovered schemas
+        self.persist_states()?;
+
+        Ok(())
+    }
+
+    /// Helper method to load schemas from a specific directory
+    pub(crate) fn load_schemas_from_directory(
+        &self,
+        dir: &PathBuf,
+        states: &HashMap<String, SchemaState>,
+    ) -> Result<(), SchemaError> {
+        if let Ok(entries) = std::fs::read_dir(dir) {
+            for entry in entries.flatten() {
+                let path = entry.path();
+                if path.extension().map(|e| e == "json").unwrap_or(false) {
+                    if let Ok(contents) = std::fs::read_to_string(&path) {
+                        let mut schema_opt = serde_json::from_str::<Schema>(&contents).ok();
+                        if schema_opt.is_none() {
+                            if let Ok(json_schema) =
+                                serde_json::from_str::<JsonSchemaDefinition>(&contents)
+                            {
+                                if let Ok(schema) = self.interpret_schema(json_schema) {
+                                    schema_opt = Some(schema);
+                                }
+                            }
+                        }
+                        if let Some(mut schema) = schema_opt {
+                            self.fix_transform_outputs(&mut schema);
+                            let name = schema.name.clone();
+                            let state = states.get(&name).copied().unwrap_or(SchemaState::Available);
+                            {
+                                let mut available = self.available.lock().map_err(|_| {
+                                    SchemaError::InvalidData(
+                                        "Failed to acquire schema lock".to_string(),
+                                    )
+                                })?;
+                                available.insert(name.clone(), (schema.clone(), state));
+                            }
+                            if state == SchemaState::Approved {
+                                let mut loaded = self.schemas.lock().map_err(|_| {
+                                    SchemaError::InvalidData(
+                                        "Failed to acquire schema lock".to_string(),
+                                    )
+                                })?;
+                                loaded.insert(name.clone(), schema);
+                                drop(loaded); // Release the lock before calling map_fields
+
+                                // Ensure fields have proper ARefs assigned
+                                let _ = self.map_fields(&name);
+                            }
+                            info!(
+                                "Loaded schema '{}' from {} with state: {:?}",
+                                name,
+                                dir.display(),
+                                state
+                            );
+                        }
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// Loads schema states from sled and loads schemas that are marked as Approved.
+    pub(crate) fn load_schema_states_from_disk(&self) -> Result<(), SchemaError> {
+        let states = self.load_states();
+        info!("Loading schema states from sled: {:?}", states);
+        info!(
+            "DEBUG: load_schema_states_from_disk called with {} states",
+            states.len()
+        );
+        let mut available = self
+            .available
+            .lock()
+            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
+        let mut schemas = self
+            .schemas
+            .lock()
+            .map_err(|_| SchemaError::InvalidData("Failed to acquire schema lock".to_string()))?;
+
+        for (name, state) in states {
+            info!("DEBUG: Processing schema '{}' with state {:?}", name, state);
+            if state == SchemaState::Approved {
+                // Load the actual schema from sled database into active memory
+                match self.db_ops.get_schema(&name) {
+                    Ok(Some(mut schema)) => {
+                        info!(
+                            "Auto-loading approved schema '{}' from sled with {} fields: {:?}",
+                            name,
+                            schema.fields.len(),
+                            schema.fields.keys().collect::<Vec<_>>()
+                        );
+
+                        // 🔄 Log ref_atom_uuid values during schema loading
+                        info!(
+                            "🔄 SCHEMA_LOAD - Loading schema '{}' with {} fields",
+                            name,
+                            schema.fields.len()
+                        );
+                        for (field_name, field_def) in &schema.fields {
+                            use crate::schema::types::Field;
+                            match field_def.ref_atom_uuid() {
+                                Some(uuid) => info!(
+                                    "📋 Field {}.{} has ref_atom_uuid: {}",
+                                    name, field_name, uuid
+                                ),
+                                None => info!(
+                                    "📋 Field {}.{} has ref_atom_uuid: None",
+                                    name, field_name
+                                ),
+                            }
+                        }
+
+                        self.fix_transform_outputs(&mut schema);
+                        info!(
+                            "After fix_transform_outputs, auto-loaded schema '{}' has {} fields: {:?}",
+                            name,
+                            schema.fields.len(),
+                            schema.fields.keys().collect::<Vec<_>>()
+                        );
+                        schemas.insert(name.clone(), schema.clone());
+                        available.insert(name.clone(), (schema, state));
+                        drop(schemas); // Release the lock before calling map_fields
+                        drop(available); // Release the lock before calling map_fields
+
+                        // Ensure fields have proper ARefs assigned
+                        let _ = self.map_fields(&name);
+
+                        // Re-acquire locks for the next iteration
+                        available = self.available.lock().map_err(|_| {
+                            SchemaError::InvalidData("Failed to acquire schema lock".to_string())
+                        })?;
+                        schemas = self.schemas.lock().map_err(|_| {
+                            SchemaError::InvalidData("Failed to acquire schema lock".to_string())
+                        })?;
+                    }
+                    Ok(None) => {
+                        info!("Schema '{}' not found in sled, creating empty schema", name);
+                        available.insert(name.clone(), (Schema::new(name), SchemaState::Available));
+                    }
+                    Err(e) => {
+                        info!("Failed to load schema '{}' from sled: {}", name, e);
+                        available.insert(name.clone(), (Schema::new(name), SchemaState::Available));
+                    }
+                }
+            } else {
+                // Load the actual schema from sled for non-Approved states too
+                match self.db_ops.get_schema(&name) {
+                    Ok(Some(mut schema)) => {
+                        // 🔄 Log ref_atom_uuid values during schema loading (non-Approved)
+                        info!(
+                            "🔄 SCHEMA_LOAD - Loading schema '{}' (state: {:?}) with {} fields",
+                            name,
+                            state,
+                            schema.fields.len()
+                        );
+                        for (field_name, field_def) in &schema.fields {
+                            use crate::schema::types::Field;
+                            match field_def.ref_atom_uuid() {
+                                Some(uuid) => info!(
+                                    "📋 Field {}.{} has ref_atom_uuid: {}",
+                                    name, field_name, uuid
+                                ),
+                                None => info!(
+                                    "📋 Field {}.{} has ref_atom_uuid: None",
+                                    name, field_name
+                                ),
+                            }
+                        }
+
+                        self.fix_transform_outputs(&mut schema);
+                        info!(
+                            "Loading schema '{}' from sled with state {:?} and {} fields: {:?}",
+                            name,
+                            state,
+                            schema.fields.len(),
+                            schema.fields.keys().collect::<Vec<_>>()
+                        );
+                        available.insert(name.clone(), (schema, state));
+                    }
+                    Ok(None) => {
+                        info!("Schema '{}' not found in sled, creating empty schema", name);
+                        available.insert(name.clone(), (Schema::new(name), state));
+                    }
+                    Err(e) => {
+                        info!(
+                            "Failed to load schema '{}' from sled: {}, creating empty schema",
+                            name, e
+                        );
+                        available.insert(name.clone(), (Schema::new(name), state));
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+}
+
